+++ finagle-core/src/main/scala/com/twitter/finagle/ssl/session/UsingSslSessionInfo.scala
@@ -5,6 +5,10 @@ import com.twitter.util.security.ProxyX509Certificate
 import com.twitter.util.security.NullSslSession
 import java.security.cert.X509Certificate
 import javax.net.ssl.SSLSession
+import org.bouncycastle.asn1.x509.Extension
+import org.bouncycastle.asn1.x509.GeneralName
+import org.bouncycastle.asn1.x509.GeneralNames
+import org.bouncycastle.cert.X509CertificateHolder
 
 /**
  * Class which indicates that a particular connection is using SSL/TLS,
@@ -12,7 +16,9 @@ import javax.net.ssl.SSLSession
  *
  * @param session The `SSLSession` associated with the connection.
  */
-private[finagle] final class UsingSslSessionInfo(val session: SSLSession) extends SslSessionInfo {
+private[finagle] class UsingSslSessionInfo(
+  val session: SSLSession)
+    extends SslSessionInfo {
 
   // This class should not be used with a `NullSslSession`. Use
   // `NullSslSessionInfo` instead.
@@ -29,8 +35,7 @@ private[finagle] final class UsingSslSessionInfo(val session: SSLSession) extend
    * The Session ID associated with an `SSLSession`.
    *
    * @note The maximum length for an SSL/TLS Session ID is 32 bytes. This method
-   * returns a hex string version of the Session ID which has a maximum length of 64 bytes.
-   *
+   *       returns a hex string version of the Session ID which has a maximum length of 64 bytes.
    * @return a hex string version of the raw byte Session ID.
    */
   val sessionId: String = session.getId.hexlify
@@ -46,9 +51,8 @@ private[finagle] final class UsingSslSessionInfo(val session: SSLSession) extend
    * The `X509Certificate`s that were sent to the peer during the SSL/TLS handshake.
    *
    * @note If certificates are `Certificate` values but not `X509Certificate` values,
-   * they will not be returned via this field. Instead use the `SSLSession#getLocalCertificates`
-   * method to retrieve those local certificates.
-   *
+   *       they will not be returned via this field. Instead use the `SSLSession#getLocalCertificates`
+   *       method to retrieve those local certificates.
    * @return The sequence of local certificates sent.
    */
   val localCertificates: Seq[X509Certificate] = {
@@ -59,6 +63,17 @@ private[finagle] final class UsingSslSessionInfo(val session: SSLSession) extend
         case cert: X509Certificate =>
           new SystemIdentityCertificate(cert)
       }
+  }
+
+  /**
+   * Private helper to extract the general names from the SubjectAlternativeName x509 extension field of a cert.
+   */
+  private def getSans(certificate: X509Certificate): Seq[GeneralName] = {
+    val holder = new X509CertificateHolder(certificate.getEncoded)
+    Option(
+      GeneralNames
+        .fromExtensions(holder.getExtensions, Extension.subjectAlternativeName))
+      .map(_.getNames.toSeq).getOrElse(Nil)
   }
 
   /**
@@ -76,6 +91,10 @@ private[finagle] final class UsingSslSessionInfo(val session: SSLSession) extend
         new SystemIdentityCertificate(cert)
     }
 
+  override def getLocalIdentity: Option[ServiceIdentity] =
+    localCertificates.headOption.map(getSans).flatMap(ServiceIdentity.apply)
+  override def getPeerIdentity: Option[ServiceIdentity] =
+    peerCertificates.headOption.map(getSans).flatMap(ServiceIdentity.apply)
 }
 
 private[session] class SystemIdentityCertificate(underlying: X509Certificate)